v1 complete. 12 vidoes processed

+++ processor_process_script.py
@@ -2,6 +2,7 @@ from processor import Processor
 import openai
 import pandas as pd
 from utils import is_whitespace_or_punctuation
+import os
 
 class ProcessorProcessScript(Processor):
     def get_name(self):
@@ -17,9 +18,10 @@ class ProcessorProcessScript(Processor):
         return ".txt"
 
     
-    def clean_find(self, text, sub_string):
+    def find_last_para_position(self, text, sub_string):
         if not sub_string:
             return len(text)
+        
         cleaned_sub_string = [ch for ch in sub_string if not is_whitespace_or_punctuation(ch)]
         cleaned_sub_string = cleaned_sub_string[:10]
         j = len(cleaned_sub_string) - 1  
@@ -43,18 +45,34 @@ class ProcessorProcessScript(Processor):
             txt = file1.read()
         chunk_index = 0
         chunk_size = 2048
-        file2 = open( script_processed_file_name, 'w')
+#        file2 = open( script_processed_file_name, 'w')
+        if os.path.exists(script_processed_file_name):
+            with open(script_processed_file_name, 'r') as file2:
+                processed_text = file2.read()
+                i_left = processed_text.rfind('[') 
+                i_right = processed_text[i_left:].rfind(']')
+                last_idx = processed_text[i_left:i_left + i_right+1]
+                chunk_index = txt.rfind(last_idx) + len(last_idx) + 1
+            file2 = open(script_processed_file_name, 'a')
+        else:
+            file2 = open(script_processed_file_name, 'w')
+
+
         while chunk_index < len(txt):
             print(f'processing {script_file_name} {chunk_index}')
             chunk = txt[chunk_index: chunk_index + chunk_size]
             processed_chunk = self.process_text_chunk(chunk)  
 
             if chunk_index + chunk_size < len(txt):
-                last_para_indx = processed_chunk.rfind('\n') 
-                last_para = processed_chunk[last_para_indx+1:]
-                in_chunk_indx = self.clean_find(chunk, last_para) 
-                processed_chunk_cut = processed_chunk[:last_para_indx]
-                chunk_index += in_chunk_indx
+                last_para_indx = processed_chunk.rfind('\n')                 
+                last_index_right = processed_chunk[:last_para_indx].rfind(']')
+                last_index_left = processed_chunk[:last_index_right].rfind('[')
+                last_idx = processed_chunk[last_index_left:last_index_right+1]
+                pos = txt.rfind(last_idx)
+                if pos == -1:
+                    pos = 0
+                chunk_index = pos + len(last_idx) + 1
+                processed_chunk_cut = processed_chunk[:last_para_indx+1]
             else:
                 chunk_index = len(txt)
                 processed_chunk_cut = processed_chunk
@@ -63,9 +81,10 @@ class ProcessorProcessScript(Processor):
         file2.close()
     
     def process_text_chunk(self, chunk):
+
         response = openai.chat.completions.create(
-                model="gpt-4",
-                messages=[{"role": "user", "content": f"给下面基督教牧师的讲道加标点，分段落.其他不要改动\n\n{chunk}"}],
+                model="gpt-4-turbo",
+                messages=[{"role": "user", "content": f"给下面基督教牧师的讲道加标点，分段落.其他不要改动,,保留Markdown Tag\n\n{chunk}"}],
                 temperature=0.5,
                 max_tokens=4000,
                 top_p=1.0,
